apps/api/requirements.txt
# Django Ninja API Requirements

# Core Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-ninja>=1.0,<2.0

# Database
psycopg2-binary>=2.9,<3.0
django-postgres-vector>=0.2,<0.3

# Authentication
PyJWT>=2.8,<3.0
python-jose>=3.3,<4.0
django-cors-headers>=4.3,<5.0

# Redis & Celery
django-redis>=5.4,<6.0
celery>=5.3,<6.0
redis>=5.0,<6.0

# Real-time
centrifugo>=5.0,<6.0

# Validation
pydantic>=2.5,<3.0
email-validator>=2.1,<3.0

# Payment Gateways (Nigerian)
requests>=2.31,<3.0

# Utilities
python-dotenv>=1.0,<2.0
Pillow>=10.0,<11.0
reportlab>=4.0,<5.0
qrcode>=7.4,<8.0

# Development
pytest>=8.0,<9.0
pytest-django>=4.7,<5.0
black>=24.0,<25.0
flake8>=7.0,<8.0
isort>=5.13,<6.0